helix: make movement commands in normal mode never select
Previously, doing `e` in normal mode would select everything until the
end of the current word. This differed from `e` in select mode in only
one way: the select mode bind would keep the selection on subsequent
key presses, while in normal mode, it would immediately forget the
previous selection.

Now, `w` and `e` don't select at all. This gives them a clearer
difference from select mode. `w` works with BEGINNINGS, while `e` works
with ENDS.

`e` is the most useful for going forward, since it will go
until the end of the current word. This matches the behavior seen in
Ctrl+→ in typical text editors. It's less useful going backwards, since
it matches the *end* of the previous word.

`w` is less useful for going forward, where it jumps to the beginning of
the next word. However, when going backward, it jumps to the beginning of
the current word. This matches the behavior of Ctrl+←.

These commands each have a variant with WORDS instead of words, triggered
using Alt.

In the future, we may change the way these are split. Rather than having
`e` be good for going forward, and `W` be good for going backward, we
could have the core functionality integrated into one key. But that's for
a later commit.