helix: add keybind to run python files
I'm tempted to wait for #61 to implement this, given the amount of work
that's gone into it, but I can't reject a genuinely easy solution. I
still believe that doing #61 will be better in the long term. That's
definitely coping, but let me cope, okay? Just check the date on this
commit message and you'll understand.