helix: make `e` the primary bind for moving around
Previously, we inherited the paradigm from Helix (which inherited it from
Neovim, which inherited it from Vim) that `w` was for moving with the
beginnings of words, and `e` was for moving with the ends of words. I
extended this paradigm, by moving the `b` keybind to `W`, since it was
still dealing with the beginning of a word -- just the previous one rather
than the next one. However, this led to a weird setup. `e` and `W` were
now my most used binds, since they corresponded to what I had the muscle
memory for: Ctrl+Right-Arrow and Ctrl+Left-Arrow.

Now, we've unified these most useful binds into one. `e` goes to the end
of the current word, while `E` goes to the beginning of the current word.
`w` does the same as always, with ends of words. We've binded `W` to
nothing, since nobody really needs to go to the end of the previous word.
Even Helix didn't have a bind for this, so why should I?

From base Helix keybinds, the changes now look like this:

`e` -> `e` (goto end of current/next word)
`e` -> `Alt-e` (goto end of current/next WORD)
`b` -> `E` (goto beginning of current/previous word)
`B` -> `Alt-E` (goto beginning of current/previous WORD)

`w` -> `w` (goto beginning of next word)
`W` -> `Alt-w` (goto beginning of next WORD)

`b` -> `no_op`