helix: swap to nixd instead of nil
Stil have things to get working when it comes to getting the version
numbers to popup, seeing if I can decrease memory usage, and getting the
`hm` alias to work. But this is a good place to start.

Do note - things weren't working for ages because the Home-Manager module
for Helix doesn't check that all set options are valid at build-time, and
I believe this goes doubly for language-server setup. This meant all the
times I was trying to configure nixd options, it wasn't in the right
location. If you're trying to configure nixd within Helix, ensure your
config goes within `language-server.nixd.config.nixd`, and then within
`options` if you're setting the expr options.