helix: stop providing pylsp its own package instance
Previously, we gave Helix access to pylsp and all its plugins by creating
another instance of Python with Pylsp and its plugins as that instance's
"modules" (what you can think of as the PIP requirements). But, this meant
that Helix was using a different instance of Python than the rest of my
system. This was bad for many reasons.

Now, we directly override pylsp to add our plugins as dependencies. This
is unfortunately a lot more verbose already, and we haven't replicated the
custom plugins yet. But it's far better in practice.