helix: go back to movement commands also selecting
We had previously gotten rid of selection here in 721eb04. However, as
I've been trying to get more practice with word-based movement, this has
actually become an inconvenience. It turns what would be `ec` into `vec`.

To bring this back, we get rid of doAndDeselect, since we're only using
it once now, and can just inline the functionality. We also bring back
the paradigm of a function accepting a `type` parameter.