helix: standardize selection collapsing when inserting
This is a nice balance between constant selection when we don't want it,
and making the movement binds not select at all.

Now, when we do `e`, it'll select the word. You can do `a` to insert at
the end of that word, and `i` to insert where you started. You could also
 do `c` or `d`, but let's ignore those for now.

The big idea is that when you do `a` or `i`, we shouldn't keep the
selection. This was frustrating and was the underlying reason I made the
harsh change of removing selection on movement binds entirely. Now, the
selection makes it easy to do something like `c` or `d`, but doesn't get
in your way for insertion.