helix: disable SC2164 via bash-language-server
After a lot of searching Github for the exact attribute to use for
accessing bash-language-server config, I finally found an example of
someone just using `environment` to set the option, which actually worked.
I don't know whether to blame home-manager or Helix for the lack of build
errors when it comes to language-server config. Regardless, it's awful
UX, and often just "guess the attribute name, rebuild, retry".