helix: pin to older version of nixpkgs
At some point between the last rev I had of nixpkgs and the current one,
something changed with tree-sitter that made helix fail to build. I'm
not sure when this was, and need to do further research for a possible
bug fix. But, for now, let's just pin the working version.