helix: make `x` always select and never extend
We already were correct with `X`, it was just the `x` bind that was wrong.
Now, `x` just changes the selection status at the cursor's line. If we
select upwards via `X`, then want to remove some selection via `x`, we
can do it now, and it won't start selecting on the bottom.