git, zsh-abbr: add `pbranch` alias and make `gbr` point to it
`pbranch` just adds formatting to `git branch` typical output, to make it
easier to chose which branch should be switched to. While it's trivial to
add an echo statement before the output, we need to make sure that as
`gbr`, the abbreviation should handle any command expected of `git branch`.
So, sadly, echo statements are out the window. Still great, though, and
in the future we can add color to make it even better.