helix: automatically trim selections for (almost) all movement binds
I just learned that `_` in helix trims leading and trailing whitespace
from your selection. Wonderful! The goal was to apply this to all
navigation binds automatically, but unfortunately, this broke with `w`.
it would get stuck on the end of a word, move one character to the right
past a space, then trim and go back to the end of that word. So, we just
trim everything except `w`.

There's a Helix PR that adds dedicated movement for selecting words
without any whitespace. I'm currently working on getting the Helix
package overriden so I can add in arbitrary PRs. Once I get that working,
I can rely on that instead.