git: add `pswitch` to switch branches with FZF
Fun story, the Regex wasn't even the worst part of this. It was the fact
that the Nix repl, in its infinite wisdom, shows an escaped backslash
symbol as two backslashes. This put me in an infinite hell when trying
to represent the string `\|`. The repl would show `\|` as `|`, and `\\|`
as `\\|`. After much searching, I finally learned that `\\|` is how the
repl represents the string `\|`.

However, I did learn about a better way to do what I was doing, that's
less broken. Instead of assigning to a variable and printing the value
of the variable, just do `:p "myStringContents\|"` in the repl, and it'll
"preview" the string as you'd expect. Useful!