zsh-abbr: add non-patch abbrs for (un)staging
Now, we have two sets of staging binds. `hire`, `fire`, and `kill` have
non-patch twins in `add`, `unstage`, and `goback`. Separating `unstage`
and `goback` is useful for a cleaner history, and also makes
better abbreviations.

Each of these new binds also has a mass-variant, for dealing with the
entire tree. This is honestly the main purpose of these new binds, since
if I wanted to stage/unstage/remove every change, I had to type lots of
stuff.

Along with all this is a change to shorten the `git reset` aliases from
`grst` to `grs`. I would split it out into another commit, but hunk
splitting is hell.