WIP
Still doesn't work with `mkUnfreeNixpkgs`, because we need to find a
clean way to get an instance of `pkgs` to merge with `utils` in the
flake. The cleanest way to do this might be to make `forAllSystems` no
longer a myLib function, and just have it within the `flake.nix`. Then,
we can use it real quick for getting a `pkgs` instance to combine when
declaring `utils`.

Maybe there's also some lazy magic we could do here? Not sure.