llakaLib: access impure version
Upstream (me, I'm upstream in this case) made it so llakaLib has two
outputs (three outputs technically, but who's counting) - the pure
version, accessed via `pureLib`, and the impure version, accessed via
`fullLib`. The impure version gives you access to functions like
`writeFishApplication` that require something from `pkgs` to work.
We actually have to instantiate both types - the pure one so we can use
`llakaLib` functions in the flake when we don't have access to system,
and the impure one to pass into `packages` and `nixosConfigurations`.