apps: move all folders into core
The new paradigm:

Ideally, folders act as a binary bucket. The optimal split is something
like `cli` and `gui`: if you're looking for an application, you know
which way to turn. The split between `apps` and `config` is messier:
many things are found in `features` that would be expected to be found
in `dev`. The recent changes to the folder structure are meant to
clarify these cases of confusion.

To explain the structure, we must first ignore `core`. With three `apps`
folders, the distinctions are mostly clear. `gui` is for GUI
applications, `extras` is for helper programs defining functionality,
and `dev` is for configuration related to the development experience.
The distinction between `extras` and `dev` is best seen with some
examples. `extras` would hold Ripgrep, `dev` would hold Direnv. `extras`
would hold Bat, `dev` would hold Python. Stuff that isn't "fun", but
functional, goes in `dev`.

Now, we reach the distinction between `core` and `extras`. This is
necessarily unclear. Anything that lives in `core` USED to live in
`extras`, but it's become integral enough to the configuration that it
had to be split into multiple files. This means, if something gets a
folder, it should no longer live in `extras`. `dev` does NOT have this
limitation, because things can't graduate from `dev` to `core`. To put
it simply, `core` is where `extras` go to grow up.

The largest source of confusion in the existing configuration is what
it means for something to go in `config/features`. Readline, for
example, was moved from `config/features` to `apps/dev`. Features was
a relatively clear distinction when the `apps` folder didn't exist, but
now anything in `apps` that touches the system could be a candidate for
`features`, and vice versa. Nix registry configuration, or TMPFS
configuration, clearly go in Features. But many things are not so
simple. More work should be done in this area.