fish: ignore instances of `fg` and `bg`
I've tried with no luck to make the bind simply execute `fg` without
putting it on the commandline. No matter what I do, it seems to leave
bad data on the fish title. We instead compromise and just ensure `fg`
is never added to history. The only problem with the current
implementation now is that the previous commands get cluttered by
instances of `fg`, but I'm not sure if there's actually a way to avoid
this without taking the strong stance of constantly running `clear`.