features: init
We provide a new module `features`, which is simply used for getters and
setters. Whenever we define something that's used across the
configuration, we define it as an option in `features` and set it
locally. Then, wherever we rely on that feature, we grab the value from
`config`, and assert that it's what we expected it to be. This means
that if we ever remove that feature, or change which program provides
it, we'll get custom errors wherever it's used, and can track down +
remove that program state.