flake: update flake.lock
We pin a specific version of home-manager unstable, because a recent
home-manager commit started using new lib functions for shell
integration functions, which breaks things when we override to use
unstable. I could try making a PR to backport these functions, but it
might get rejected. Might as well try.