kittab: init
Fixes #74. We create a custom desktop entry `kittab.desktop`, and use it
for everything that `kitty.desktop` would previously be used for.
Everything going through it means that anything calling Kitty will
instead create a new tab, as long as an instance of Kittab already
exists.

To ensure we don't get issues in the future, we add our first boolean
assertion, that kittab exists and is set up. It's better than storing
`kittab` to `features.terminal`, since that's far more unclear.