yazi: use generic cwd function and use when closed
I was basically writing the exact same function every time I wanted to
write the CWD on doing an action. Instead, I just made a Nix function
that writes the directory and does some Yazi action. We use this to also
write the directory when pressing escape, since that's nice for the
workflow of changing directory in yazi.nvim, pressing escape,
suspending, pressing Ctrl+y, and then running some command.