yazi/escexit: init
The previous changes adding the `modified` plugin made me realize that
I'd actually gotten rid of the `escape` behavior with Yazi, so I could
never stop filtering. We now make <Esc> first check if there's any
pickers or filters to be closed. If so, it does that. If not, it closes
Yazi and stores the cwd. I realize while writing this that maybe I could
get away with using mkCwd somehow, making it a native Yazi plugin rather
than a simple Nix function, but this is fine for the time being.
Although, that makes me realize that could be a way of running commands
at Yazi runtime... hmm, maybe this deserves a Github issue.