kittab: go back to using normal icon
I was *wrong* in the previous commit. The icon doesn't change whether
kitty is in the nix store: it's kitty itself. Running `echo $PATH` will
show Kitty only if that command is being run through Kitty. Why? No
clue. Waiting for a friend to reproduce the issue. Maybe it's time for a
nixpkgs issue.