yazi: use yaziPlugins from nixpkgs
Thanks to khaneliman, we don't have to fetch from the repo anymore! This
is especially nice because the repo sometimes bumped the required
version before nixpkgs - now, that shouldn't be an issue anymore (given
that the plugins get updated, of course - but I hope that won't be a
problem). Also comes with a flake input bump since I needed to grab the
plugins from the latest `nixpkgs-unstable`, and thought I might as well
bump everything.