modules/baseVars: remove `isDefined` checks
This was added when I knew far less about the module system. I was
running into issues where I wanted to ensure that the value was set, but
found that without this, it wouldn't be checked. Turns out, it *is*
checked without `isDefined` - just only if the value is actually used.
That's why I wasn't getting the error before - I hadn't used it yet.
Makes sense!