taskbar: abstract into a feature
There's a desktop in the works that needs its settings set through
something that isn't dconf, so we need access to the taskbar apps from
somewhere that can be accessed in multiple places. We can't use
`baseVars`, since they're designed to not have access to `config`, which
we need for the assertions. So, instead, we make a new feature!