flake: update to unstable early
Had the time and decided I might as well. We keep `nixpkgs-unstable` and
`pkgs-unstable` - they just point to the same inputs right now as
`nixpkgs` and `pkgs`. When 25.05 comes out in a week or two, we'll swap
off. We also move a lot of packages off unstable that only needed it for
something temporary - the ones that actually prefer the unstable life
get to stay.