portals: use gtk by default, add kde portal
We always want the gtk portal, just in case. But, we only want the Gnome
and KDE terminals when we're actually using that desktop. Maybe in the
future we'll have them always on without mkIf - because having the gnome
one off might make occasional changes to gnome not work very well. Just
depends on whether having both applied causes fights between them or
not.