git: revamp abbrs
All the git abbrs are used a TON. To solve this, we shorten them to not
use the `g` prefix. This makes a lot of the ones I use most one-letter,
which is wonderful. Will have to get used to these, but I think it's
good in the long run.

We also get rid of a few abbrs that I honestly don't use - these include
`gfe` for `git fetch`, `gcm` for `git commit --message`, `gamp` for `git
amend --patch`, etc. The only one sticking around that I don't really
use is `ac` for `git commit .`, but I want to start using that one,
since it brings `a<CR>c<CR>` into a single command.

And, finally, a few aliases are gone that I don't use. The idea for `git
pbranch` was kinda neat, but I think having it more deterministic is
better. I also probably haven't used the stash abbreviations in at least
six months.