baseVars, hostVars: pass through specialArgs
I want to do things in a way that's ubiquitous, and specialArgs is a
great way of injecting global constant variables. If you need something
that can be modified anywhere, a module is your best friend - but
`baseVars` and `hostVars` will always be defined toplevel, so it makes
sense to just pass them directly.