outputs: split nioxsConfigurations into toplevel file
Previously, all the hostVars were in a subdirectory, while the mkNixos
declaration stuff was still toplevel. This wasn't very helpful, since
the two are coupled together. Now, they share a file in `hosts.nix`.
This makes `outputs.nix` smaller and easier to work with, while keeping
direct host logic front and center. We also rewrite basically all the
comments to be even clearer.