structure: rename toplevel directories
`apps` -> `programs`
`config` -> `system`
`other` -> `various`

As always, I want the folders I interact with most to be on top. That
means we need to enforce alphabetical order for our directories, which
limits us a bit. Ideally, I would keep the misc directory as `other`,
but then it ends up first in the directory list, and I can't have that.
I'm okay with having a subpar name for my misc folder, if it makes the
folders I actually interact with better.