myLib/recursivelyImport: run toString on path
Huge performance improvement. If you send a path into hasSuffix, it'll
copy it to the store, which we can avoid by using `toString` beforehand.

Before this patch, it took about 2.54 seconds to get all the nix files
in nixpkgs - now it takes 0.42 seconds. Of course, this is for nixpkgs,
which has 80 thousand files. For my repo, this saved 30 milliseconds