fish: revamp keybinds
The day of the vi-mode is over. I find that for interfaces like a
cmdline, it's more effective to always be in insert mode, and have
control motions for basic movement operations. We now have `ctrl-[hjkl]`
binds, with `ctrl-l` and `ctrl-j` taking double duty as tab-complete and
execute. Finally, we make the control arrow binds move by a full WORD,
which I find more intuitive. I probobly won't remember to use the shift
ones, but what can you do.