wrappers: init
If the future is wrappers, then I'd be giving up the ability to "query"
things about packages. This is a proof-of-concept for how to keep that
without any dependencies - using only some fancy functions. Each
"wrapper" declares a module and a package. The package does all the
necessary wrapping, and results in a derivation. The module provides
some option interfaces for other packages to query. We can even load the
modules automatically, thanks to `listFilesRecursive`.

Not sure if this will be the impl we'll stick with - I'm proud of it in
the moment, but I have to give it some time to see if I'm actually
willing to commit to it.