nixos: call `mkHost` directly, redefine hostname
The previous solution is very beautiful, but it's a lot less
approachable. Since I like my config to be a decent reference, I'd
prefer to simplify, even if it ends up leading to a bit more verbosity.
Rather than running `mapAttrs` on the whole attrset, and assuming the
hostname from the keys, we just redefine it in `hostVars`. Unnecessary?
Yes! But now it's easier to tell what's going on. We also reorganize
each host's `hostVars` section to follow a simple template (required
vars on top, optional vars on the bottom).