wrappers/git: init
Lots of stuff coming with this! The `gh` and `less` modules both get
`iniConfig` options, so we don't have to hardcode their logic within
`git/settings.nix`. We're merging these together manually, as Adios
isn't like the module systems you know - merges are manual!

We also add a diff-so-fancy module. I had a lot of strange issues with
getting diff-so-fancy to work properly. It seemed like no matter what I
did, it would color added/removed newlines in white. Finally, I tracked
down the issue - it seems to read from the GIT_CONFIG environment
variable, to determine your git settings. I didn't want to add this
environment variable globally, so I ended up giving it its own wrapper.
I felt it was also a little gross to have the module just do
`${inputs.git.drv}/git/config`. Instead, I have a simple linkFarm
derivation that contains the git config, which the git drv and the
diff-so-fancy drv both depend on.

However, because the diff-so-fancy module now depended on the git
module, I ended up getting infinite recursion when I tried to move the
diff-so-fancy iniContent to the diff-so-fancy module. I asked for a fix
for this on the adios side, since it SHOULD be possible to have two
modules depend on each other, as long as they use different options. For
now, I hardcode the ini content on the git side, and just install
`diff-so-fancy` systemwide.

Other future work includes wiring up my custom packages to be able to
access wrappers. satod should be able to rely on the custom
diff-so-fancy package, rather than just assuming it exists systemwide.