various: don't use `programs.fish.functions`
There's no equivalent to this outside of home-manager - and most of the
time, I should really just put things in `interactiveShellInit` There
are cases where something really needs to be in `functions/` -
down_or_execute is a function that `bind` needs to be aware of). But for
that, we just do a manual symlink.

Indentation beware! I ran into lots of issues writing this, that seemed
to originate from `hm.programs.fish.functions` running `fish_indent` on
the original text. I thought this was unnecessary, but without
four-space indentation, down_or_execute refused to work. Word of warning
if you end up going on a similar path.