wrappers/firefox: use json files where possible
Trying to make my wrappers simply wrap "normal" code, and only embed
things in Nix when actually necessary. I've taken that to its extreme
here, to the extent of using `replaceVars` so I don't need any nix files
for configuration at all. `wrapFirefox` is horribly documented, and this
was hell.