wrappers/starship: increase truncation length, substitute `~`
I wanted to allow showing longer paths, but I'm actually a fan of just
showing Documents/projects/foo most of the time. The solution? Increase
the truncation length, but substitute out ~/ so it never appears. I can
revert this later if I change my mind, but I like it so far.