starship: consistently fix double space problems
I want to have two separate git sections in the statusline - one for
staging, one for pushing. Having `+1` along with the staging stuff is
gross to me. However, I'd like a space between the two if they're both
there, so you get something like:

[S] [+1]

But the space _doesn't_ appear if there's only one or neither.

At first glance, you'll find Starship makes this difficult. It does have
a way to only show something if the variables are non-empty, by wrapping
the content in parentheses. But it works like an `any` function, not an
`all` function. However, this is solvable - we just wrap each part in
its own parentheses, and have those parts append a trailing space.
Genius!