programs/fish: move to `fish_user_key_bindings` function
Removing preset binds wasn't working when it was in the generic initrc.
This is going to cause an issue later, when I want to stop using hm for
fish (and eventually transition it into a wrapper). The nixos module
doesn't read from `/etc/fish/functions`, so I need to make sure the dir
is actually vendored. Problem for a future me.