npins: update sources
Now using unstable neovim. In theory we use the same commit as in my
nvim flake, so we shouldn't have to build separately here - but I want
to use the npins derivation fetcher so we're not victim to the
tarball-ttl cache. This means that the input ends up hashing slightly
differently, and we build separately. Maybe I should go flakeless on my
neovim config, or move it into this repo - always decisions to make.